// client/src/constants/endPoints.ts
export enum mainRoutes {
  admin = '/admin',
  auth = '/auth',
  courses = '/courses',
  coursepage = '/coursepage',
  coursemanager = '/coursemanager',
}

export enum endPoints {
  all = '/all',
  course = '/course',
  lesson = '/lesson',
  subjects = '/subjects',
  subscribe = '/subscribe',
  unsubscribe = '/unsubscribe',
  subscriptions = '/subscriptions',
  created = '/created'
}